requirements.txt
fastapi
uvicorn
python-multipart
nltk
spacy
scikit-learn
sentence-transformers
numpy
pandas
pydantic
python-docx
PyPDF2
fpdf
python-dotenv
textstat
pyspellchecker
groq

googlesearch-python
requests
beautifulsoup4
